bilbo and Mummi contribution!! veriverse contribution!! https://LiveEdu.tv/ (Livecoding.tv) contribution!! ThomasBS (Thomas Boregaard Sørensen, https://kaem.dk/, https://kaem.io/, tbhs1000%💛forgrowing!! Den bedste udvikler!), drmjg, camwhite (Cameron White, Twitchs' scamwhite), xmetrix, intelijens, syaseigirl, CatherinaHelena (CatherinaCortega, Twitchs' helenawinter), fuel, r00tkiller, hiiamgosu (itashellhaha...(gosu), scorpioinyouranus... ), pablosj (sjpablo, Pawel Kulesza, Pablosj special contribution (after so much days with so low audience... ) .. contribution!!), phanxgames, 0din, 7ru57, anita, beater, benjones, bill_gates, brayoeh, camilosasuke (me :ca2:), cary, clementin, Cookiesliyr, costinis, crisy, dauntlesstech, david_joffe, denvercowboy, drcode, dslegends, ederangelo21..., epryka, evgeniy_andreev, famed, flythiago, frankly444, georgeb, gigaherz, gostrafx, grahambell, guacjack, hatyme, hbw, heinz3, hishritaha, hosjenken, intelijens, irwanbd, isabella, jackperrie, jacobs, jaycee, johnnydully, kaeris, kamisama, kannu, karrmarr, kayne3, kissloryshy, klerind, kozmo, kranalex, kruzoe, landry96, lemonay, lex_yang, larrybrook, LiveEdu_bot, liveedu_support, LiveOrDevTrying, lyndroid, maoabc1818, mark_miller, masterchoc, matroskinb, mitch_ill, mitsurugi, nero951, ngetich, novicehere, olivert, olelukoe, ong, pestka12, pinkgorilla, pinky, pirateofthesea, polodonald, rex2, resident, riverside, rooney2, samdev, sargin, shgalym (from Kazakhstan), shieldgenerator7 (Twitchtv shieldgenerator7), shinkaisokuuindo, siox, starlord91, stefan_v, stephen22, systemis, teleporter, thermonucleary, thiagovgamg, tonycode, valdinei, vapaamies, vodkavodkavodka228, wariosolis, wetriks, wilson2000, yanaseattle, zakaria98, zzeroge and zzzel contribution!! https://Twitch.tv/ contribution!! ! ThomasBS_ (LiveEdu.tv ThomasBS, Den besten Software udvikler), RocketBeans.TV (Nils, Simon, Colin, Donni, Etienne, Florentin, Lars, Marco, Oliver, Renè, [Schröck, Daniel] und Viet), AlfredoPlays (and his mates... Alfredo, JackieMButler, Hanson, Rey and Mackie contribution), KindaFunnyGames (Greg Miller, Nick Scarpino, Andy and Kevin contribution...), SoulFoamTV (LettuceSmokeStudios, SoulFoam), scamwhite (Cameron White, Liveedutvs' camwhite, mesh.io), Phanxgames, Peter_Alfons_Loch (from phanxgames streamd and discord channel, first "explicit" and "spontaneous" commit and contribution to this newly create GitHub repository), cancerMoonLife(hiiamgosu from LiveEdu), Spoofh (LiveEdu.tv Fuel: Benjamin Kuhl), themushmouth (LiveEdu.tv mushmouth: Patrick McFarling), Tectuma (Kevin, Stacie (GothicCat1), Matt Brock (Cronos) and Dylan and Nick...), IILomoo (Liveedu lomoo, Matti), nothings2 (Sean Barret), MitchJones (Nick [Nicholas], Greek, sodappoping contribution), sonnystar (hairy green sonny star, Vlad Ionescu), WildN00b (Wild from LiveEdu/LiveCoding), gosu (hiamgosu xxxKLEOPATRAxxx contribution), RobTheSwan (and Teazal cat contribution... ), helenawinter (CatherinaHelena, CatherinaCortega from LiveEdu), XanderGamingTwich and xmxstudio (LiveCoding/LiveEdu xmetrix), Adan_204, Adevel, Always_Betonentei, Cylverrose, david_joffe, dimer525, Dreamdischarger, dzyrr_tv, emielrregis, faillehad (french swearer with igrectepe) ), Fataldizz, foldm, Funmartas_cz, GameDuells, Getlostdude, Grononmc, guy_ross, Hazurl, igrectepe (French swearer with faillehad) , itechno2016 (Russian guy), Javo_net, knutaf, Kubule, Lauheid1974, Lessio__, LiveOrDevTrying, LoryMaster, Luxio_tv, MatejKunc, Miouchfr, Mopynatv, OrcaPrime, OxenHill, Piepassaro, Pixel701, Pucki4, recursionishard, Restreamiobot. shieldgenerator7 (Liveedutv shieldgenerator7), SystemOfPeter, Tanerax, theformemrcc, Tookietookie, Trump88, Volon_, Vulsen, Wraithone,  contribution!! https://mixer.com/ (previous Beam.pro) contribution!! R4ver contribution (and his friends agile as him, from Sverige... contribution), ArrestedMoney2, LlamaJuic3467, NormalZeus53319, sharkshooter641, TaxingBinkie154 and WedRhombus23 contribution!! YouTube Live contribution!! NzLeoGenX X, Shivam Bhola contribution!! bisba contribution!! Carol, Lúcio and Lucinho contribution!! Upwork.com contribution!! Bock Engineering contribution!! Infinisys and Dennis Heazle contribution!!
[Sep2017] 2nd Windows Store App Round
from bilbo at veriverse!! Internet Net Curitiba 60Mbps▼ 6Mbps▴
i7 6700 Dell XPS 8900 32GB DDR4 RAM 2x240GB SSD 2TB HDD
Windows 10 Pro (Quad Boot with Windows 10 Pro with Hyper-V, Windows 8.1 Pro, macOS Sierra 10.12.6 and Ubuntu 17.04)

+++ appseed/aura/aura/os/metrowin/metrowin_small_ipca_channel.cpp
@@ -78,7 +78,7 @@ namespace aura
 
          return true;
 
-   }
+      }
 
 
 #endif
@@ -106,7 +106,7 @@ namespace aura
 
          Uri ^uri = ref new Uri(anotherappURI);
 
-         ::wait(Launcher::LaunchUriAsync(uri), durationTimeout.get_total_milliseconds());
+         ::wait(Launcher::LaunchUriAsync(uri), (DWORD) durationTimeout.get_total_milliseconds());
 
          return true;
 
@@ -132,7 +132,7 @@ namespace aura
 
          memory m;
 
-         
+
 
          string anotherappURI = m_strBaseChannel + "://send?messagebin=" + ::str::from(message)+ "," + ::aura::system::g_p->url_encode(::aura::system::g_p->base64().encode((byte *)pdata, len));
 
@@ -271,7 +271,7 @@ namespace aura
 
       }
 
-      
+
       ipc::ipc(::aura::application * papp) :
          object(papp),
          base(papp),
